Clean up OSGINorthbound.newNettyEndpointProps()
We have a magic check for URI scheme -- i.e. if we grow a different
transport, we might happily be using 'http' without ever realizing it.

Refactor logic to make it absolutely clear which cases we handle, so
that there can be no mistakes.

Also clean up obvious gaps in semantics of objects we create and add
FIXMEs/TODOs for follow-ups.

Change-Id: I5a5b812d137fc4d665cfa457a1df1d5bbe4968e6
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>