Add reproducer for NETCONF-1386
When the pattern includes an element that allows repetitions and it's
preferred over the next one because the former has the lower value,
then it is repeated as many times as allowed.
This causes StackOverflowError if the number of repetitions is very
large or unlimited (e.g. '*').

The first test works because '[0-9]' is preferred over '[A-Z]' but
when it's reversed in the 2nd one, test fails with StackOverflowError.
The 3rd one also fails for the same reason - '[a-zA-Z0-9_.-]' is
preferred over '[a-zA-Z0-9])' hence PropertyEntity.prepareExample
is stuck in the infinite loop without making progress beyond the
element.

JIRA: NETCONF-1386
Change-Id: Ib7b7714330b5176541fe928dc5660c20579f8b2d
Signed-off-by: Sangwook Ha <sangwook.ha@verizon.com>