Rework resource lookup
RestconfRequestDispatcher is using indirection via PathParameters to
interpret the requested path.

There are multiple problems here:
- the logic is quite undocumented
- there is an overload of PathArguments.EMPTY to report at least two
  distinct paths
- we are reporting 409 where we should be reporting 404, as noted in
  RFC8040 section 4.3
- the splitting works on decoded path, whereas the ApiPath.parse()
  should always we presented with an encoded path

This patch deals with the first three problems, documenting how we are
peeling the path and what gaps there are relative to the specification.

The last problem is marked with a FIXME and will be addressed in a
follow-up patch.

JIRA: NETCONF-1379
Change-Id: I408626a5e328bb938f727d9a549ca8297c9df3a9
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>