Reconstruct Target URI, part one
In order to correctly serve our requests, we need to properly
reconstruct target URI, as per HTTP/1.1 spec.

This patch deals with the simplest case, asterisk-form: it is only
allowed for OPTIONS request and reports all methods implemented by
the server. All other methods fails with either status 400.

That necessitates knowledge as to what methods are requested, hence
we have a statically-populated Set.

JIRA: NETCONF-1379
Change-Id: Id37044799460759d1066ad7a17524ee0c0284060
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>