Reconstruct Target URI, part two
Reconstructing Target URI from its usual origin-form is a tad more
involved:

1. we need access to the URI scheme being used, which we now have
   conveniently available from HTTPTransportChannel
2. we need to ensure we have a Host header, that it contains only one
   value and that it is a valid authority

We employ java.net.URI for most of the parsing, combining and error
reporting functions. It also allows us to determine in the uri is in
absolute-form, in which case we can pick it up as is.

We use this URI to initialize QueryStringDecoder, so that its uri()
method reports the complete thing.

JIRA: NETCONF-1379
Change-Id: I6641a613f2fd4c63332ae29b0a876ca32a325a03
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>