Fix request path interpretation
We are decoding the request path before processing it, which leads to
our inability to properly target values.

This patch introduces SegmentPeeler, which aids our traversal of the
path -- thus we can examine decoded segments while we still retain the
ability to pick up the remainder of the path in encoded form.

JIRA: NETCONF-1399
Change-Id: Ib7f7b08db8a97da7e9f60528f9834c50e6ce1ee4
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>