Remove NettyEndpointConfiguration.baseUri()
The idea that the user needs to configure a hostname for Netty endpoint
is quite wrong: we get an absolute URI with each request and hence can
trivially derive restconfURI that the caller can use to reach us.

Remove the host-configuration knob and streamline
NettyEndpointConfiguration by propagating only the restconf() prefix.

JIRA: NETCONF-1379
Change-Id: Ia4857e74e109e04339865fec31f4261343ab953d
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>