Update Netty RESTCONF API root path configuration
The 'restconf' configuration property defaults to "rests" for historical
reasons and controls the {+restconf} expansion on the JAX-RS endpoint.

In the Netty endpoint we want to retain the ability to control the path,
but want to start with the default of "restconf", to make the
out-of-the-box experience match RFC8040.

This patch introduces a separate "api-root-path" property, with a
well-documented syntax and semantics.

We this property via NettyEndpointConfiguration in a decoded form, so
there is no confusion as to what the actual semantics are -- which is
always a problem with naked Strings.

JIRA: NETCONF-1379
Change-Id: I14ae912082c9407cac796971686e75f0c23cf04f
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>