Mark outbound write thread is internal
When we are writing out a message, we may be hitting a KEX -- at which
point we would normally be blocked.

The problem is that the blocked thread is usually the Netty IO thread,
which effectively ends up deadlocking it.

Prevent this situation by marking our thread as internal, which causes
SSHD to enqueue the message rather than blocking.

JIRA: NETCONF-1415
Change-Id: If472986270e827a316099168069754fe1661939d
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>
(cherry picked from commit 8f5cd19b584e047522a93acb12e594a5cb08a5cc)