Add WebHostResource
We need the ability to inject secondary services into the NettyEndpoint,
such that they have a reference to the XRD (for now).

This patch introduces WebHostResource as the common execution API, which
are dispatched to from EndpointRoot.

Lifecycle is integrated with OSGi SR, but other implementations should
be straightforward to add.

JIRA: NETCONF-1370
Change-Id: I041059919db7471a11c5e0da75557ff3c48efcda
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>