Add WebHostResource
We need the ability to inject secondary services into the NettyEndpoint,
such that they have a reference to the XRD (for now).

This patch introduces WebHostResource as the common execution API, which
are dispatched to from EndpointRoot.

This maps neatly to OSGi whiteboard pattern:
- WebHostResourceProvider is the service published to the whiteboard
- OSGiNettyEndpoint instances are event sources

In other environments SimpleNettyEndpoint's explicit register()/close()
mechanics might prove to be helpful

JIRA: NETCONF-1370
Change-Id: I041059919db7471a11c5e0da75557ff3c48efcda
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>