Use virtual threads for request execution
HTTPServerSession has four distinct phases, two of which involve
reading or writing message body.

This patch offloads processing from the Netty thread, so that the task
of parsing request body and wrapping of asynchronous execution occurs
on a dedicated virtual thread -- off-loading IO threads and allowing
Channel.write() to occur in a blocking context.

JIRA: NETCONF-1424
Change-Id: Ia347e673d824ca18374a390284095410b361c5a4
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>