Notify transport channel on session close
Call notifyTransportChannelFailed method when SshClient
or SshServer invokes sessionClosed.
If ChannelInactive close the session and TransportChannel
is not ready, then it is necessary to call
notifyTransportChannelFailed which triggers the reconnection process.

This needs to be handled here, because the NetconfClientSession
lifecycle has not started yet. The NetconfClientSession is created
in NetconfNodeHandler only when ClientTransportChannelListener is
successful. If ClientTransportChannelListener is successful and session
is closed, then ConnectingTask will never finish nor provide any
additional information.

JIRA: NETCONF-1423
Change-Id: Ief230b2d28a747628488d0dba3700fbb552d163d
Signed-off-by: Peter Suna <peter.suna@pantheon.tech>
(cherry picked from commit 72a6440320fb5174ce9b9026311c01e25f7950b3)