Switch over to HTTPScheme-driven channel setup
Setting up the entire pipeline in HTTPServer does not quite work, as we
need to know what HTTP version we are going to use to properly setup
execution.

This patch shifts the setup from HTTPServer to
HTTPServerSession.handlerAdded(), whence we defer to HTTPScheme set up
the pipeline before HTTPServerSession.

JIRA: NETCONF-1419
Change-Id: I6eaeb753b04aac57a6f81190a1fde535534197ec
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>
Signed-off-by: Ivan Hrasko <ivan.hrasko@pantheon.tech>