Refactor HTTPScheme.initializeServerPipeline()
The split nature of setup makes it quite unclear as to what is going on.

This improves the code flow, as we now understand that channelHandler()
is a HTTP2->HTTP1 thing -- hence we can allocate it only when needed.
This leads to TLS-based-HTTP1 not allocating it at all.

JIRA: NETCONF-1419
Change-Id: Ib8998b648fac27536c2b7b64fac6818de1fb0374
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>
(cherry picked from commit a41052500628ada8ab2f0ef0d8897f339f91a6a4)