Split out CleartextUpgradeHandler
The code structure is clear: we have two handlers:
- AlpnUpgradeHandler handles the transition from TLS for HTTPS
- CleartextUpgradeHandler handles the transition frop TCP for HTTP

Split out the anonymous class into a named entity, so we have an
improved debugging experience -- including transition debug messages.

We also talk to our superclass to not automatically release the message,
and therefore we do not need to undo its work in channelRead().

JIRA: NETCONF-1419
Change-Id: I428e1ed91d7c3c5a6beb51af56fd8abcac112a16
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>