Migrate HttpClientServerTest
Use HTTPServerSession for test harness, as we are entering a phase where
mocking would have to duplicate production code.

Since we are using HTTPServerSession, we can verify our task gets
executed on the corresponding virtual thread. That in turn allows us to
ditch the executor service in favor of an explicit sleep.

JIRA: NETCONF-1419
Change-Id: If1958fd6e57180565cd3d0bd5b341cfa0427b3b9
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>