Stop edit-config chain on error
Currently, the code executes edit-config operations sequentially,
continuing execution even if one operation fails. This results in all
edit-config operations completing, regardless of individual errors.

This commit modifies the chaining of edit-config operations into
the resultFeature list. The result of each operation is now checked for
errors. If an error is found, an TransactionEditConfigFailedException
is created, halting the execution chain.

JIRA: NETCONF-1197
Change-Id: Ie38e6eff920eb0931ae30e7469b657bd8e515d1a
Signed-off-by: Peter Suna <peter.suna@pantheon.tech>