Clean up features
Factoring things out has moved our dependencies a bit, making
odl-netconf-util and odl-netconf-ssh dead features.

This renames odl-netconf-util to odl-netconf-mdsal-common, carrying a
dependency on mdsal-dom-api. We package netconf-mdsal-common and
netconf-dom-api here.

We then drop netconf-auth from odl-netconf-api, as it really is a thing
specific to the auth plugin.

JIRA: NETCONF-714
Change-Id: I3267a73ee8e8f8c32a9764a6a642aa17264b6fa0
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>