Split out restconf-server-api-testlib
The use of test-jar is rather ugly, as we cannot control package a
proper module-info specific to tests. This leads to server.api being
exposed from its module, but also from the unnamed-module.

Split out the small library we have into a separate artifact, which
solves these problems.

It also provides a basic prototype of a JPMS-enabled binding packaging,
albeit it does not currently work with binding-data-codec.

Change-Id: I59d21cd779a216911418807e0d9c31bf0558032f
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>