Promote DatabindPath/ServerErrorPath
DatabindPath is a very useful construct, as it provides the context in
which a YANG databind operation is performed. Promote it to
netconf-common for wider reuse.

ServerErrorPath is a common modeling construct similar to ErrorMessage:
it is shared between NETCONF and RESTCONF as the value of error-path
element. Promote it as o.o.netconf.common.ErrorPath.

Change-Id: I279ef4b64f471d43f573e1825dbc38a37fe81f87
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>