Introduce NetconfServerStrategy
Now that we have DatabindContext available, we can make it the base of
our data binding state to an EffectiveModelContext.

A NetconfServerStrategy is a combination of DatabindContext and
capabilities derived from its EffectiveModelContext -- resulting in
proper caching of computed capabilities and better code locality.

All users of CurrentSchemaContext.getCurrentContext() can be migrated to
DatabindProvider.currentDatabind(), so we do that as well, improving
modularity.

This leads to lower need to mock and allows us to move the extension
wiring to be something internal to CurrentSchemaContext.

Change-Id: Ie43fd18e9ba8432627423d0c0adc4f0d551daf55
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>