Verify correct value type in YangPatch body
If the provided JSON input is not a string and is recognized
as BEGIN_ARRAY, BEGIN_OBJECT, or NULL, calling in.nextString()
will produce an IllegalStateException, like:
IllegalStateException: Expected a string but was BEGIN_ARRAY

Verify that the provided JSON input is a string. If the input
is not a string, raise a RequestException.

JIRA: NETCONF-1438
Change-Id: Ica4a9ab93c3fcff85842eceaa6e9bbe9c81b15ef
Signed-off-by: Peter Suna <peter.suna@pantheon.tech>