Handle empty 'value' in YangPatch
If the 'value' node contains empty brackets ("value": {}) in
the provided JSON, it throws an IllegalStateException, and restconf
replies with a 500 Server Error.

In the case of XML, when <value></value> is provided, XmlPatchBody
produces an IndexOutOfBoundsException, and restconf replies with
a 500 Server Error.

Handle this case and reply with error message
"Empty 'value' element is not allowed"

JIRA: NETCONF-1438
Change-Id: I4377c2db9050b8594c42992c1197d919a0ba4e5e
Signed-off-by: Peter Suna <peter.suna@pantheon.tech>