Store correct Subscription in StreamRegistry
Currently, we are loosing result of createSubscription(subscription)
and storing only SubscriptionImpl in subscriptions map inside
StreamRegistry. This leads to lookupSubscription() returning only
SubscriptionImpl and for example termination() is only invoked on
SubscriptionImpl and MdsalRestconfStreamSubscription is never
terminated as we don't have access to it, this leads to subscription
not being deleted from datastore.

JIRA: NETCONF-714
Change-Id: I4bb59db9cb432e851f7dd5f372685f708e1b03ff
Signed-off-by: Samuel Schneider <samuel.schneider@pantheon.tech>