Remove redundant non null check in JsonPatchBody
Remove the redundant check for a non-null TargetSchemaNode
in JsonPatchBody. Reason:

- The TargetSchemaNode is not used inside this method or under
this if block.
- If the target is not null, then TargetSchemaNode should also
not be null, or it should fail here as an invalid state:
JsonPatchBody.PatchEdit#setTargetSchemaNode()
- The only place where TargetSchemaNode is used is here:
JsonPatchBody#readEditDefinition()
so in this state, it is already checked.

JIRA: NETCONF-1438
Change-Id: Ia08bcbbeb14f425d0b0a430b21257fa6fcf7bffe
Signed-off-by: Peter Suna <peter.suna@pantheon.tech>
(cherry picked from commit a3342a0cbb35994b224715c39ad25fab5bfedfc8)