Adapt subscription termination on session closed
Logic to remove subscription requires request that initiated removing,
but when subscription remove is invoked by channel being inactive,
there is no request.
Separated method is created for cleaning of subscriptions after the
session is closed. Logic from terminateImpl extracted to separated
method with nullable request, so it can be used for both cases.
Removed sending of subscriptionTerminated as there is no one to send
it to.

JIRA: NETCONF-714
Change-Id: Icc237700900be286cf20134abd52c730ff4a769f
Signed-off-by: Matej Sramcik <matej.sramcik@pantheon.tech>