Isolate 'filterSpec == null' case
When we have a filter without a spec we should call removeFilter(), not
try to putFilter(null), which will induce a NPE because we try to put it
into a null-hostile map.

This is a known invariant, so cover it with nullness annotations, which
point out problems which will be addressed in subsequent patches.

JIRA: NETCONF-1459
Change-Id: I1e409464810256a2839532caba552c8026a272fd
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>