Improve filter population scalability
The DTCL we register here is effectively a DataListener on the entire
filters map node and update all filters present in it.

This provides O(N) scalability where the N is the number of existing
filters -- i.e. in the user updates a single filter, we end up updating
them all.

We only need two things from the filter: its name and its filter-spec,
so a much more efficient way is to use a wildcard to listen to
filter-specs of all filter.

The exercies of updating filters then only has to deal with filter specs
coming and going, eliminating one level of NormalizedNode lookups.

The filter name is part of the path to the filter-spec, so we can
recover it from the supplied YangInstanceIdentifier.

JIRA: NETCONF-1459
Change-Id: I08832709e7d8b81797aa6a30f914c9174248e811
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>