Stop subscription on stop-time reached
ScheduledExecutorService is used to create a task that will terminate
subscription after stop-time is reached. Added stop-time task
termination to other methods of subscription termination so they
will not be left running.
After stop-time is reached, subscription-terminated is sent instead of
subscription-completed as that is marked as "if-feature: configured" in
a model, what we don't support.

JIRA: NETCONF-1466
Change-Id: I5835f53762fa5eb51e228177b2fdbeba505c7eb5
Signed-off-by: Matej Sramcik <matej.sramcik@pantheon.tech>