Support both 2011 and 2013 base-module revisions
When a device advertises only the 2013-09-29 revision of
urn:ietf:params:xml:ns:netconf:base:1.0, the connector failed on the
second (schema-aware) GET/GET-CONFIG request because the hard-coded
QNames still referenced the 2011-06-01 revision that is no longer part
of the EffectiveModelContext.

This change remaps the SourceIdentifier used for
EffectiveModelContext creation to tolerate the 2011/2013 revision
mismatch and prefer the device-advertised revision where appropriate.
Register a built-in ietf-netconf@2011-06-01 YANG into the device
SharedSchemaRepository before schema resolution. We load the resource
from the model bundle via YangModuleInfoImpl and register it as
a local YangTextSource.

JIRA: NETCONF-1474
Change-Id: I244420595896d99770637c97980ab618e44bcc79
Signed-off-by: Yaroslav Lastivka <yaroslav.lastivka@pantheon.tech>
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>