Support both 2011 and 2013 base-module revisions
When a device advertises only the 2013-09-29 revision of
urn:ietf:params:xml:ns:netconf:base:1.0, the connector failed on the
second (schema-aware) GET/GET-CONFIG request because the hard-coded
QNames still referenced the 2011-06-01 revision that is no longer part
of the EffectiveModelContext.

NetconfDeviceRpc#invokeNetconf now checks whether the
QName’s module is present in the current schema if not,
rewrites the request QName to the latest revision
before dispatching it to NetconfDeviceDOMRpcService.

JIRA: NETCONF-1474
Change-Id: I244420595896d99770637c97980ab618e44bcc79
Signed-off-by: Yaroslav Lastivka <yaroslav.lastivka@pantheon.tech>