Remove CandidateWithRunning implementation
The combination of Running and Candidate data store is redundant and
complicates the lock operation.

The CandidateWithRunning implementation performs Candidate operations
with one exception. It also locks the Running data store. The lock
of Running data store is not necessary when other operation is only
on Candidate data store.

JIRA: NETCONF-1352
Change-Id: I963713a127b119f40effdc7b119ad838daedc774
Signed-off-by: Peter Suna <peter.suna@pantheon.tech>