Add concurent-rpc-timeout-millis config
If concurrent-rpc-limit is set, any request that is unable to acquire
a semphore lock automaticaly fails. This creates situatuion where it
is not possible to initialize connection with concurrent-rpc-limit=1.

To resolve this, the concurrent-rpc-timeout-millis setting has been
introduced. This allows a user to specify a wait time for
a semaphore lock.

JIRA: NETCONF-1498
Change-Id: I253fcfdb16bb96b9ffc1b5c8fcf26ba657cd1f77
Signed-off-by: Peter Suna <peter.suna@pantheon.tech>