Remove unrequired lock for pollRequest method
The java.util.concurrent.Semaphore does not need to be under
ReentrantLock and can handle concurret enviroment on its own.
Removing this allows for releasing the semaphore without being
blocked while another thread tries to acquire semaphore thread.

Using peek to check if we can use poll is not efficient
and requires a lock for both operations. A more efficient solution
is to use just poll, which returns null when the queue is empty,
with a concurrent implementation of Queue.

JIRA: NETCONF-1498
Change-Id: I2cc60e9e45ebeb87d1b84ab45f4dadc6a93d0437
Signed-off-by: Peter Suna <peter.suna@pantheon.tech>