Ignore absent stop-time on modify-subscription
When stop-time field is not present in modifySubscriptionRpc, the value
of stop-time in subscription and datastore is left unchanged.
Null values are not supported in RPCs, so null only indicates the field
was not provided, in which case we do not update it.

JIRA: NETCONF-1491
Change-Id: Id5304f8079cf5cf43a2818889b4092a9523944f7
Signed-off-by: Matej Sramcik <matej.sramcik@pantheon.tech>
(cherry picked from commit f7663a25ef0a8d261a3080cdf4f5a621668e791a)